@@ -1,132 +1,125 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
-import 'package:http/http.dart' as http;
-import 'package:path/path.dart' as path;
-import 'package:path_provider/path_provider.dart';
-import 'package:uuid/uuid.dart';
 
 void main() {
-  runApp(const ImageDownloaderApp());
+  runApp(const MyApp());
 }
 
-class ImageDownloaderApp extends StatelessWidget {
-  const ImageDownloaderApp({Key? key}) : super(key: key);
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
 
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Image Downloader',
+      title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
+        // This is the theme of your application.
+        //
+        // TRY THIS: Try running your application with "flutter run". You'll see
+        // the application has a purple toolbar. Then, without quitting the app,
+        // try changing the seedColor in the colorScheme below to Colors.green
+        // and then invoke "hot reload" (save your changes or press the "hot
+        // reload" button in a Flutter-supported IDE, or press "r" if you used
+        // the command line to start the app).
+        //
+        // Notice that the counter didn't reset back to zero; the application
+        // state is not lost during the reload. To reset the state, use hot
+        // restart instead.
+        //
+        // This works for code too, not just values: Most code changes can be
+        // tested with just a hot reload.
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
       ),
-      home: const ImageDownloaderPage(title: 'Image Downloader Home Page'),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
-class ImageDownloaderPage extends StatefulWidget {
-  const ImageDownloaderPage({Key? key, required this.title}) : super(key: key);
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+
+  // This widget is the home page of your application. It is stateful, meaning
+  // that it has a State object (defined below) that contains fields that affect
+  // how it looks.
+
+  // This class is the configuration for the state. It holds the values (in this
+  // case the title) provided by the parent (in this case the App widget) and
+  // used by the build method of the State. Fields in a Widget subclass are
+  // always marked "final".
 
   final String title;
 
   @override
-  _ImageDownloaderPageState createState() => _ImageDownloaderPageState();
+  State<MyHomePage> createState() => _MyHomePageState();
 }
 
-class _ImageDownloaderPageState extends State<ImageDownloaderPage> {
-  static const imagesDir = 'images';
-  final Uuid uuidGenerator = const Uuid();
-  late Future<List<File>> _imageFiles;
+class _MyHomePageState extends State<MyHomePage> {
+  int _counter = 0;
 
-  ScrollController? _scrollController;
-
-  @override
-  void initState() {
-    super.initState();
-    _scrollController = ScrollController(
-      initialScrollOffset: 0.0,
-      keepScrollOffset: false,
-    );
-    _imageFiles = _loadImages();
-  }
-
-  Future<Directory> _getImagesDirectory() async {
-    final directory = await getApplicationDocumentsDirectory();
-    return Directory(path.join(directory.path, imagesDir)).create();
-  }
-
-  Future<void> _downloadNewImage() async {
-    final response = await http.get(Uri.parse("https://picsum.photos/400"));
-    final imagesDir = await _getImagesDirectory();
-    final filePath = path.join(
-      imagesDir.path,
-      "${uuidGenerator.v4().toString()}.jpg",
-    );
-    final file = File(filePath);
-    await file.writeAsBytes(response.bodyBytes);
-    print("Image downloaded to: $filePath");
-
+  void _incrementCounter() {
     setState(() {
-      _imageFiles = _loadImages();
+      // This call to setState tells the Flutter framework that something has
+      // changed in this State, which causes it to rerun the build method below
+      // so that the display can reflect the updated values. If we changed
+      // _counter without calling setState(), then the build method would not be
+      // called again, and so nothing would appear to happen.
+      _counter++;
     });
-  }
-
-  Future<List<File>> _loadImages() async {
-    final imagesDir = await _getImagesDirectory();
-    return imagesDir
-        .list()
-        .where((element) => element.path.endsWith(".jpg"))
-        .map((e) => File(e.path))
-        .toList();
-  }
-
-  void _scrollToEnd() {
-    if (_scrollController != null) {
-      _scrollController!.animateTo(
-        _scrollController!.position.maxScrollExtent,
-        duration: const Duration(milliseconds: 1000),
-        curve: Curves.bounceOut,
-      );
-    }
   }
 
   @override
   Widget build(BuildContext context) {
+    // This method is rerun every time setState is called, for instance as done
+    // by the _incrementCounter method above.
+    //
+    // The Flutter framework has been optimized to make rerunning build methods
+    // fast, so that you can just rebuild anything that needs updating rather
+    // than having to individually change instances of widgets.
     return Scaffold(
       appBar: AppBar(
+        // TRY THIS: Try changing the color here to a specific color (to
+        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
+        // change color while the other colors stay the same.
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        // Here we take the value from the MyHomePage object that was created by
+        // the App.build method, and use it to set our appbar title.
         title: Text(widget.title),
       ),
-      body: FutureBuilder<List<File>>(
-          future: _imageFiles,
-          initialData: const [],
-          builder: (context, snapshot) {
-            SchedulerBinding.instance?.addPostFrameCallback((_) {
-              _scrollToEnd();
-            });
-            if (snapshot.hasData) {
-              final imageFiles = snapshot.data;
-              return GridView.builder(
-                controller: _scrollController,
-                itemCount: imageFiles!.length,
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                    crossAxisCount: 2),
-                itemBuilder: (context, index) {
-                  return Image.file(imageFiles[index]);
-                },
-              );
-            }
-            return const Center(
-              child: CircularProgressIndicator(),
-            );
-          }),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _downloadNewImage,
-        tooltip: 'Download image',
-        child: const Icon(Icons.file_download),
+      body: Center(
+        // Center is a layout widget. It takes a single child and positions it
+        // in the middle of the parent.
+        child: Column(
+          // Column is also a layout widget. It takes a list of children and
+          // arranges them vertically. By default, it sizes itself to fit its
+          // children horizontally, and tries to be as tall as its parent.
+          //
+          // Column has various properties to control how it sizes itself and
+          // how it positions its children. Here we use mainAxisAlignment to
+          // center the children vertically; the main axis here is the vertical
+          // axis because Columns are vertical (the cross axis would be
+          // horizontal).
+          //
+          // TRY THIS: Invoke "debug painting" (choose the "Toggle Debug Paint"
+          // action in the IDE, or press "p" in the console), to see the
+          // wireframe for each widget.
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'You have pushed the button this many times:',
+            ),
+            Text(
+              '$_counter',
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
+          ],
+        ),
       ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _incrementCounter,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }
